Update README.md
I wanted to get the word out about a set of Playgrounds I created that graphically show how to use machine learning algorithms.  Currently I have ones for linear regression, support vector machines, and neural networks.

I put the proposed addition in 'Libraries', as they reference the AIToolbox framework, although all the code needed is in the Source directory of the Playground - so they do not require a framework to be installed.  They would also fit in the 'Algorithms' section.

The link is also to the Playground subdirectory of the AIToolbox repository, rather than the AIToolbox main page

Thanks for your consideration!